% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Table.R
\name{Create_Table}
\alias{Create_Table}
\title{A Create_Table Function}
\usage{
Create_Table(
  SCdata,
  BeforeFilter,
  mincells = 3,
  mingenes = 500,
  mtpercent = 20,
  rbpercent = 50,
  nRows = 40,
  coreFont = 2.5,
  colheadFont = 2.5,
  rowheadFont = 2.5,
  titlesize = 25
)
}
\arguments{
\item{SCdata}{Seurat object to count number of cells.}

\item{mincells}{Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{mingenes}{Include cells where at least this many features are detected.}

\item{mtpercent}{Include cells reporting at most this much mitochondrial transcript percentage.}

\item{rbpercent}{Include cells reporting at most this much ribosomal transcript percentage.}

\item{nRows}{Maximum number of rows to be prined.}

\item{coreFont}{core Font Size.}

\item{colheadFont}{colhead Font Size.}

\item{rowheadFont}{rowhead Font Size.}

\item{titlesize}{titlesize Font Size.}
}
\description{
This function create table for Seurat Normalzied data.
}
\examples{
Create_Table()
}
\keyword{Sample}
\keyword{matrix.DIR,}
\keyword{mincells,}
\keyword{mingenes,}
\keyword{mtpercent,}
\keyword{rbpercent,}
