% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet_Detection_DF_DD.R
\name{Doublet_Detection_DF_DD}
\alias{Doublet_Detection_DF_DD}
\title{A Doublet_Detection_DF_DD Function}
\usage{
Doublet_Detection_DF_DD(
  SeuratObject,
  saveDIR,
  Sample,
  Species = "hsa",
  FeatureUseCount = 2500,
  PCAnum = 10,
  resClus = 0.5,
  ClusPallette = ClusPallette,
  DoubletFinder = TRUE,
  DoubletDecon = TRUE,
  plotCCgene = TRUE
)
}
\arguments{
\item{SeuratObject}{Seurat Object.}

\item{saveDIR}{Path to save Quality plots and RDS data.}

\item{Sample}{Sample Name.}

\item{Species}{Species Name. Valid options are hsa or mmu.}

\item{FeatureUseCount}{Number of features to select as top variable features; only used when selection.method is set to 'dispersion' or 'vst'}

\item{PCAnum}{Number of PCs to be used}

\item{resClus}{Resolution to be used for clustering}

\item{ClusPallette}{Color pallete to be used for clusters}

\item{DoubletFinder}{Run DoubletFinder}

\item{DoubletDecon}{Run DoubletDecon}

\item{plotCCgene}{Plots TOP2A gene or not}
}
\description{
This function allows you to detect doublets using both Doublet_Finder and Doublet_Decon algorithms.
}
\examples{
Doublet_Detection_DF_DD()
}
\keyword{ClusPallette,}
\keyword{DoubletDecon}
\keyword{DoubletFinder,}
\keyword{FeatureUseCount,}
\keyword{PCAnum,}
\keyword{Sample,}
\keyword{SeuratObject,}
\keyword{plotCCgene}
\keyword{resClus,}
\keyword{saveDIR,}
