% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet_Detection_Mouse.R
\name{Doublet_Detection_Mouse}
\alias{Doublet_Detection_Mouse}
\title{A Read10X_Norm_Mouse Function}
\usage{
Doublet_Detection_Mouse(
  SeuratObject,
  saveDIR,
  Sample,
  FeatureUseCount = 2500,
  PCAnum = 10,
  resClus = 0.5,
  ClusPallette = ClusPallette,
  DoubletFinder = TRUE,
  DoubletDecon = TRUE
)
}
\arguments{
\item{SeuratObject}{Seurat Object.}

\item{saveDIR}{Path to save Quality plots and RDS data.}

\item{Sample}{Sample Name.}

\item{FeatureUseCount}{Number of features to select as top variable features; only used when selection.method is set to 'dispersion' or 'vst'}

\item{PCAnum}{Number of PCs to be used}

\item{resClus}{Resolution to be used for clustering}

\item{ClusPallette}{Color pallete to be used for clusters}

\item{DoubletFinder}{Run DoubletFinder}

\item{DoubletDecon}{Run DoubletDecon}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Read10X_Norm_Mouse()
}
\keyword{ClusPallette,}
\keyword{DoubletDecon}
\keyword{DoubletFinder,}
\keyword{FeatureUseCount,}
\keyword{PCAnum,}
\keyword{Sample,}
\keyword{SeuratObject,}
\keyword{resClus,}
\keyword{saveDIR,}
