% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{extract_reference_profile}
\alias{extract_reference_profile}
\title{extract_reference_profile}
\usage{
extract_reference_profile(
  dat.sce.filt,
  label,
  num_signatures = 500,
  up_only = FALSE,
  rank_slot = "rank.AUC",
  lfc_slot = "median.logFC.cohen"
)
}
\arguments{
\item{lfc_slot}{}
}
\description{
extract_reference_profile
}
