% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_ChosenPCA_and_Res_Harmony.R
\name{Final_ChosenPCA_and_Res_Harmony}
\alias{Final_ChosenPCA_and_Res_Harmony}
\title{A Final_ChosenPCA_and_Res_Harmony Function}
\usage{
Final_ChosenPCA_and_Res_Harmony(
  Temp.object,
  saveDIR,
  IdentToBatchCorrect = "orig.ident",
  SuffixName = "Testing_PCA_Dim",
  ChosenPCA = 25,
  ChosenRes = 0.1,
  ColNamesToPlot = ColNamesToPlot,
  ColPaletteToPlot = ColPaletteToPlot,
  mingenes = 500,
  ClusOrder = ClusOrderFrom1
)
}
\arguments{
\item{Temp.object}{Seurat objects to be used for the QC plots.}

\item{saveDIR}{Directory to save the plots.}

\item{IdentToBatchCorrect}{Identity to be used for Harmony batch correction (Max 1)}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{ChosenPCA}{ChosenPCA}

\item{ChosenRes}{ChosenRes}

\item{ColNamesToPlot}{Vector of identities to plot (Max = 3)}

\item{ColPaletteToPlot}{list of color palettes to plot for the identities (Max = 3)}

\item{mingenes}{minimum gene number that will be mentioned in the output file name}
}
\description{
This function allows you to perform differential gene analysis.
}
\examples{
Final_ChosenPCA_and_Res_Harmony()
}
\keyword{ChosenPCA,}
\keyword{ChosenRes,}
\keyword{ColNamesToPlot,}
\keyword{ColPaletteToPlot}
\keyword{IdentToBatchCorrect,}
\keyword{SuffixName,}
\keyword{Temp.object,}
\keyword{mingenes}
\keyword{saveDIR,}
