% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Colors.R
\name{load_Colors}
\alias{load_Colors}
\title{A load_Colors Function}
\usage{
load_Colors()
}
\arguments{
\item{load_Colors}{load_Colors}
}
\description{
This function load all the required packages for Seurat Analysis.
}
\examples{
load_Colors()
}
\keyword{load_Colors}
