% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Regular_Plots.R
\name{Make_Regular_Plots}
\alias{Make_Regular_Plots}
\title{A Make_Regular_Plots Function}
\usage{
Make_Regular_Plots(
  Temp.object,
  saveDIR,
  ColNamesToPlot = ColNamesToPlot,
  ColPaletteToPlot = ColPaletteToPlot,
  SuffixName = "QC_Plots",
  Species = "hsa",
  HistogramPlot = TRUE,
  VlnPlot = TRUE,
  QCumapPlots = TRUE,
  TablePlot = TRUE,
  UMAPPlot = TRUE
)
}
\arguments{
\item{Temp.object}{Seurat objects to be used for the QC plots.}

\item{saveDIR}{Directory to save the plots.}

\item{ColNamesToPlot}{Vector of identities to plot}

\item{ColPaletteToPlot}{list of color palettes to plot for the identities}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{Species}{Species Name. Valid options are hsa or mmu}

\item{HistogramPlot}{Plot Histograms or not}

\item{VlnPlot}{Plot Violins or not}

\item{QCumapPlots}{Plot QC UMAP or not}

\item{TablePlot}{Plot Tables or not}

\item{UMAPPlot}{Plot UMAP or not}
}
\description{
This function allows you to perform differential gene analysis.
}
\examples{
Make_Regular_Plots()
}
\keyword{ColNamesToPlot,}
\keyword{ColOrdersToPlot}
\keyword{Species,}
\keyword{SuffixName,}
\keyword{Temp.object,}
\keyword{saveDIR,}
