% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_CCA.R
\name{Perform_CCA}
\alias{Perform_CCA}
\title{A Perform_CCA Function}
\usage{
Perform_CCA(
  TempAll.object,
  NameInpdf,
  saveDIR,
  FeatureUseCount = 2500,
  CCAdimchosen = 30,
  res = 0.5,
  k.filter = 200,
  plots = TRUE,
  save = TRUE
)
}
\arguments{
\item{TempAll.object}{A list of Seurat objects between which to find anchors for downstream integration.}

\item{NameInpdf}{Path to save Quality plots and RDS data.}

\item{saveDIR}{Suffix to be added to CCA files.}

\item{FeatureUseCount}{A numeric value. This will call SelectIntegrationFeatures to select the provided number of features to be used in anchor finding}

\item{CCAdimchosen}{Which dimensions to use from the CCA to specify the neighbor search space}

\item{res}{k.filter    How many neighbors (k) to use when filtering anchors}

\item{plots}{Save CCA plots}

\item{save}{Save integrated CCA RDS Seurat object}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Perform_CCA()
}
\keyword{CCAdimchosen,}
\keyword{FeatureUseCount,}
\keyword{NameInpdf,}
\keyword{TempAll.object,}
\keyword{k.filter,}
\keyword{plots,}
\keyword{res,}
\keyword{save}
\keyword{saveDIR,}
