% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_CCA_Diff_PCAdim_res.R
\name{Perform_CCA_Diff_PCAdim_res}
\alias{Perform_CCA_Diff_PCAdim_res}
\title{A Perform_CCA_Diff_PCAdim_res Function}
\usage{
Perform_CCA_Diff_PCAdim_res(
  Sampleall.object,
  NameInpdf,
  saveDIR,
  TempCCAdimchosenlist = c(15, 20, 30),
  FeatureUseCount = 2500,
  plots = TRUE,
  save = FALSE
)
}
\arguments{
\item{Sampleall.object}{A list of Seurat objects between which to find anchors for downstream integration.}

\item{NameInpdf}{Path to save Quality plots and RDS data.}

\item{saveDIR}{Suffix to be added to CCA files.}

\item{FeatureUseCount}{A numeric value. This will call SelectIntegrationFeatures to select the provided number of features to be used in anchor finding}

\item{plots}{Save CCA plots}

\item{save}{Save integrated CCA RDS Seurat object}
}
\description{
This function allows you to perform CCA at different res.
}
\examples{
Perform_CCA_Diff_PCAdim_res()
}
\keyword{FeatureUseCount,}
\keyword{NameInpdf,}
\keyword{Sampleall.object,}
\keyword{TempCCAdimchosen,}
\keyword{plots,}
\keyword{res,}
\keyword{save}
\keyword{saveDIR,}
