% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Perform_GeneOntology_Analysis_Using_Marker_Genes.R
\name{Perform_GeneOntology_Analysis_Using_Marker_Genes}
\alias{Perform_GeneOntology_Analysis_Using_Marker_Genes}
\title{A Perform_GeneOntology_Analysis_Using_Marker_Genes Function
Except EnrichR, all others uses DEGs}
\usage{
Perform_GeneOntology_Analysis_Using_Marker_Genes(
  Temp.object,
  Temp.MarkerList,
  saveDIR,
  SuffixName = "GO",
  methods = c("ClusterProfilerM", "ClusterProfilerAlt1", "ClusterProfilerAlt2"),
  EnrichDB = "GO_Biological_Process_2021",
  GroupNameDEGs = "cluster",
  OrgDB = org.Mm.eg.db,
  OrgName = "mmu",
  topnumber = 10,
  PDFheight = 12,
  PDFwidth = 18
)
}
\arguments{
\item{Temp.object}{A list of Seurat objects between which to find anchors for downstream integration.}

\item{Temp.MarkerList}{A data.frame containg marker genes.}

\item{saveDIR}{Path to save generated data.}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{methods}{ClusterProfiler (enrichGO)}

\item{EnrichDB}{EnrichR DB. For list of DBs: listEnrichrDbs()... https://maayanlab.cloud/Enrichr/#libraries}

\item{GroupNameDEGs}{ClusterProfiler (enrichGO)}

\item{OrgDB}{ClusterProfiler (enrichGO)}

\item{OrgName}{ClusterProfiler (enrichGO)}

\item{topnumber}{Max #pathways to be plotted per group}

\item{PDFheight}{PDFheight}

\item{PDFwidth}{PDFwidth}
}
\description{
This function allows you to perform differential gene analysis.
}
\examples{
Perform_GeneOntology_Analysis_Using_Marker_Genes()
}
\keyword{"",}
\keyword{=}
\keyword{PDFheight=10,}
\keyword{PDFwidth=12}
\keyword{SubmsigdbrCategorySubset}
\keyword{Temp.object,}
\keyword{minSizeGeneSet=5,}
\keyword{msigdbrCategoryList="H",}
\keyword{saveDIR,}
\keyword{topnumber=10,}
