% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_Pathway_Analysis_FGSEA.R
\name{Perform_Pathway_Analysis_FGSEA}
\alias{Perform_Pathway_Analysis_FGSEA}
\title{A Perform_Pathway_Analysis_FGSEA Function}
\usage{
Perform_Pathway_Analysis_FGSEA(
  Temp.object,
  saveDIR,
  SuffixName = "Pathway_FGSEA",
  msigdbrCategoryList = "H",
  SubmsigdbrCategorySubset = "",
  minSizeGeneSet = 5,
  topnumber = 10,
  PDFheight = 10,
  PDFwidth = 12
)
}
\arguments{
\item{Temp.object}{A list of Seurat objects between which to find anchors for downstream integration.}

\item{saveDIR}{Path to save generated data.}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{msigdbrCategoryList}{Msigdb Category List. H, C1:C8. https://www.gsea-msigdb.org/gsea/msigdb/}

\item{SubmsigdbrCategorySubset}{SubCategoryUse list based on MsigDB selected}

\item{minSizeGeneSet}{minimum Size of GeneSet for filtering}

\item{topnumber}{Max #pathways to be plotted per group}

\item{PDFheight}{PDFheight}

\item{PDFwidth}{PDFwidth}
}
\description{
This function allows you to perform differential gene analysis.
}
\examples{
Perform_Pathway_Analysis_FGSEA()
}
\keyword{"",}
\keyword{=}
\keyword{PDFheight=10,}
\keyword{PDFwidth=12}
\keyword{SubmsigdbrCategorySubset}
\keyword{Temp.object,}
\keyword{minSizeGeneSet=5,}
\keyword{msigdbrCategoryList="H",}
\keyword{saveDIR,}
\keyword{topnumber=10,}
