% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{prep_query_dat}
\alias{prep_query_dat}
\title{Title}
\usage{
prep_query_dat(dat.sce, genelist.df, gene_to_use = "hGene")
}
\arguments{
\item{dat.sce}{input query scRAseq data as SCE obj}

\item{genelist.df}{reference profile goes in here}

\item{gene_to_use}{reference profile's colname for gene names}
}
\value{
sorted query data (SCE obj) for SCA running
}
\description{
Title
}
