% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Process_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist.R
\name{Process_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist}
\alias{Process_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist}
\title{A Process_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist Function}
\usage{
Process_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist(
  Temp.object,
  FeatureNum = 2500,
  Species = "hsa",
  mt = TRUE,
  rb = TRUE,
  cc = FALSE
)
}
\arguments{
\item{Temp.object}{Seurat Object}

\item{FeatureNum}{number of Features for variable genes}

\item{Species}{Species Name. Valid options are hsa or mmu}

\item{mt}{mitochondrial genes to be removed from Seurat object}

\item{rb}{ribosomal genes to be removed from Seurat object}

\item{cc}{CellCycle genes (Converted from Human to Mouse in needed) to be removed from Seurat object}
}
\description{
This function performs a) mt, ribo \% calculation, b) Normalization c) Filtering from variable genes d) Scaling of the data
}
\examples{
Process_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist()
}
\keyword{Temp.object,}
\keyword{cc}
\keyword{mt,}
\keyword{rb,}
