% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Process_Subset_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist.R
\name{Process_Subset_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist}
\alias{Process_Subset_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist}
\title{A Process_Subset_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist Function}
\usage{
Process_Subset_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist(
  Temp.object,
  FeatureNum = 2500,
  Species = "hsa",
  mt = TRUE,
  rb = TRUE,
  cc = FALSE
)
}
\arguments{
\item{Temp.object}{Seurat Object}

\item{Species}{Species Name. Valid options are hsa or mmu}

\item{mt}{mitochondrial genes (MOUSE) to be removed from Seurat object}

\item{rb}{ribosomal genes (MOUSE) to be removed from Seurat object}

\item{cc}{CellCycle genes (Converted from Human to Mouse) to be removed from Seurat object}

\item{number}{of Features for variable genes}
}
\description{
This function allows you to filter genes for mt, ribosomal and cell cycle list.
}
\examples{
Process_Subset_Seurat_Object_Initial_Filter_Genes_from_Variable_Genelist()
}
\keyword{Temp.object,}
\keyword{cc}
\keyword{mt,}
\keyword{rb,}
