\name{sca-internal}
\alias{isblockempty}
\alias{isdiffcomp}
\alias{iswithinblock}
\alias{eqmatrix}
\alias{normmatrix}
\alias{redrawMatrix}

\title{Internal Functions of Package sca}
\description{
  Internal function of the \file{sca} package.
}
\usage{
isblockempty(P)
isdiffcomp(x)
iswithinblock(d, P)
eqmatrix(A, B)
normmatrix(P)
redrawMatrix(phase, I, Imax, P, crit, nextpc, okcomp, message, warn)
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
