% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_Cluster_Resolution_Harmony.R
\name{Test_Cluster_Resolution_Harmony}
\alias{Test_Cluster_Resolution_Harmony}
\title{A Test_Cluster_Resolution_Harmony Function}
\usage{
Test_Cluster_Resolution_Harmony(
  Temp.object,
  saveDIR,
  IdentToBatchCorrect = "orig.ident",
  SuffixName = "Testing_Cluster_Resolution",
  ColNamesToPlot = ColNamesToPlot,
  ColPaletteToPlot = ColPaletteToPlot,
  mingenes = 500,
  ClusResList = c(0.1, 0.15, 0.2, 0.3),
  ClusOrder = ClusOrderFrom1
)
}
\arguments{
\item{Temp.object}{Seurat objects to be used for the QC plots.}

\item{saveDIR}{Directory to save the plots.}

\item{IdentToBatchCorrect}{Identity to be used for Harmony batch correction (Max 1)}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{ColNamesToPlot}{Vector of identities to plot (Max = 3)}

\item{ColPaletteToPlot}{list of color palettes to plot for the identities (Max = 3)}

\item{mingenes}{minimum gene number that will be mentioned in the output file name}

\item{ClusResList}{Vector of the PCAs to be used for testing}
}
\description{
This function allows you to perform differential gene analysis.
}
\examples{
Test_Cluster_Resolution_Harmony()
}
\keyword{ColNamesToPlot,}
\keyword{ColPaletteToPlot}
\keyword{IdentToBatchCorrect,}
\keyword{PCAdimList}
\keyword{SuffixName,}
\keyword{Temp.object,}
\keyword{mingenes}
\keyword{saveDIR,}
