% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vanGalen_HSPC}
\alias{vanGalen_HSPC}
\title{data(vanGalen_HSPC)}
\format{
An object of class \code{SingleCellExperiment} with 11259 rows and 1431 columns.
}
\usage{
vanGalen_HSPC
}
\description{
data(vanGalen_HSPC)
}
\keyword{datasets}
