% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Velten_HSPC}
\alias{Velten_HSPC}
\title{Data for SCA example}
\format{
SCE objects of normal HSPC datasets
\describe {
       \item{}
       }

       @source {add source later}

       @examples
       # example code
       data(Velten_HSPC)
}
\usage{
Velten_HSPC
}
\description{
Contains the normal HSPC data from Velten et al (2017) and van Galen et al (2019)
}
\keyword{datasets}
