% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{extract_reference_profile}
\alias{extract_reference_profile}
\title{extract_reference_profile}
\usage{
extract_reference_profile(
  dat.sce.filt,
  label,
  num_signatures = 500,
  up_only = FALSE,
  rank_slot = "rank.AUC",
  lfc_slot = "median.logFC.cohen"
)
}
\arguments{
\item{dat.sce.filt}{QCed reference data as SCE obj}

\item{label}{cell type of interest to compute DEG using scoreMarkers function from scran}

\item{num_signatures}{number of}

\item{up_only}{use up-regulated genes only if prompted}

\item{rank_slot}{decide which rank slot from scoreMarkers function results}

\item{lfc_slot}{decide which lfc slot from scoreMarkers function results}
}
\value{
return reference profile with genes and lfc values from scoreMarkers function result
}
\description{
extract_reference_profile
}
