% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Genes_from_Variable_Genelist_Mouse.R
\name{Filter_Genes_from_Variable_Genelist_Mouse}
\alias{Filter_Genes_from_Variable_Genelist_Mouse}
\title{A Filter_Genes_from_Variable_Genelist_Mouse Function}
\usage{
Filter_Genes_from_Variable_Genelist_Mouse(
  SCdata,
  Species = "hsa",
  mt = TRUE,
  rb = TRUE,
  cc = FALSE
)
}
\arguments{
\item{SCdata}{Seurat Object}

\item{Species}{Species Name. Valid options are hsa or mmu}

\item{mt}{mitochondrial genes (MOUSE) to be removed from Seurat object}

\item{rb}{ribosomal genes (MOUSE) to be removed from Seurat object}

\item{cc}{CellCycle genes (Converted from Human to Mouse) to be removed from Seurat object}
}
\description{
This function allows you to filter genes for mt, ribosomal and cell cycle list.
}
\examples{
Filter_Genes_from_Variable_Genelist_Mouse()
}
\keyword{SCdata,}
\keyword{cc}
\keyword{mt,}
\keyword{rb,}
