% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_CCA.R
\name{Perform_CCA}
\alias{Perform_CCA}
\title{A Perform_CCA Function}
\usage{
Perform_CCA(
  Sampleall.object,
  FeatureUseCount = 2500,
  plots = TRUE,
  save = TRUE
)
}
\arguments{
\item{FeatureUseCount}{Number of features to select as top variable features; only used when selection.method is set to 'dispersion' or 'vst'}

\item{matrix.DIR}{Path to 10X directory.}

\item{saveDIR}{Path to save Quality plots and RDS data.}

\item{Sample}{Sample Name.}

\item{mincells}{Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{mingenes}{Include cells where at least this many features are detected.}

\item{mtpercent}{Include cells reporting at most this much mitochondrial transcript percentage.}

\item{rbpercent}{Include cells reporting at most this much ribosomal transcript percentage.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Perform_CCA()
}
\keyword{FeatureUseCount}
\keyword{Sample,}
\keyword{matrix.DIR,}
\keyword{mincells,}
\keyword{mingenes,}
\keyword{mtpercent,}
\keyword{rbpercent,}
\keyword{saveDIR,}
