% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_DGE_ONEvsALL.R
\name{Perform_DGE_ONEvsALL}
\alias{Perform_DGE_ONEvsALL}
\title{A Perform_DGE_ONEvsALL Function}
\usage{
Perform_DGE_ONEvsALL(
  Temp.object,
  saveDIR,
  SuffixName = "ALLcells",
  MainCol = "seurat_clusters",
  MainColOrder = ClusOrder,
  MainColPallete = ClusPallette,
  ColNamesToPlot = ColNamesToPlot,
  ColPaletteToPlot = ColPaletteToPlot,
  Species = "hsa",
  FDR = 0.1,
  FCcutoff = 1.5,
  topnumber = 5,
  downsampleHeatmap = 300,
  plots = TRUE,
  save = TRUE
)
}
\arguments{
\item{Temp.object}{A list of Seurat objects between which to find anchors for downstream integration.}

\item{saveDIR}{Path to save generated data.}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{MainCol}{Column to be used (Default is seurat_cluster)}

\item{MainColOrder}{Factor order to be used (Default is numerical sorted order)}

\item{MainColPallete}{color palette to be used}

\item{ColNamesToPlot}{Vector of identities to plot}

\item{ColPaletteToPlot}{list of color palettes to plot for the identities}

\item{Species}{Species Name. Valid options are hsa or mmu}

\item{FDR}{FDR cutoff}

\item{FCcutoff}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing logfc.threshold speeds up the function, but can miss weaker signals.}

\item{topnumber}{Max #genes to be plotted per identity}

\item{downsampleHeatmap}{Max #cells to be plotted per identity}

\item{plots}{Save CCA plots}

\item{save}{Save integrated CCA RDS Seurat object}
}
\description{
This function allows you to perform differential gene analysis.
}
\examples{
Perform_DGE_ONEvsALL()
}
\keyword{ColNamesToPlot,}
\keyword{ColPaletteToPlot,}
\keyword{FCcutoff,}
\keyword{FDR,}
\keyword{MainCol,}
\keyword{MainColOrder,}
\keyword{MainColPallete,}
\keyword{Species,}
\keyword{SuffixName="ALLcells",}
\keyword{Temp.object,}
\keyword{downsampleHeatmap,}
\keyword{plots,}
\keyword{save}
\keyword{saveDIR,}
\keyword{topnumber,}
