% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read10X_Norm_Mouse.R
\name{Read10X_Norm_Mouse}
\alias{Read10X_Norm_Mouse}
\title{A Read10X_Norm_Mouse Function}
\usage{
Read10X_Norm_Mouse(
  matrix.DIR,
  saveDIR,
  Sample,
  mincells = 3,
  mingenes = 500,
  mtpercent = 20,
  rbpercent = 50,
  FeatureUseCount = 2500,
  plots = TRUE,
  save = TRUE
)
}
\arguments{
\item{matrix.DIR}{Path to 10X directory.}

\item{saveDIR}{Path to save Quality plots and RDS data.}

\item{Sample}{Sample Name.}

\item{mincells}{Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{mingenes}{Include cells where at least this many features are detected.}

\item{mtpercent}{Include cells reporting at most this much mitochondrial transcript percentage.}

\item{rbpercent}{Include cells reporting at most this much ribosomal transcript percentage.}

\item{FeatureUseCount}{Number of features to select as top variable features; only used when selection.method is set to 'dispersion' or 'vst'}

\item{plots}{Save QC plots}

\item{save}{Save RDS Seurat object}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Read10X_Norm_Mouse()
}
\keyword{FeatureUseCount,}
\keyword{Sample,}
\keyword{matrix.DIR,}
\keyword{mincells,}
\keyword{mingenes,}
\keyword{mtpercent,}
\keyword{plots,}
\keyword{rbpercent,}
\keyword{save}
\keyword{saveDIR,}
