% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R, R/SCA_data.R
\docType{data}
\name{SCA}
\alias{SCA}
\title{SCA}
\usage{
SCA(
  dat.sce.filt,
  cor.method = "spearman",
  ref_profile = NULL,
  num_perm = 500,
  data.use = c("fccounts", "logcounts"),
  rand.features.use = "across.sample.refp.gene.permutation",
  ref_data_null_dist = NULL
)
}
\arguments{
\item{dat.sce.filt}{filtered SingleCellExperiment object}

\item{cor.method}{using spearman's correlation coefficient as default}

\item{ref_profile}{takes reference profile that either user provided or by extract_reference_profile function}

\item{num_perm}{set number of permutations}

\item{data.use}{which assay to use whitin SCE obj. default to logcounts}

\item{rand.features.use}{randomization method to use}

\item{ref_data_null_dist}{input background data to use. if not provided, query data will be used as a backgrouund dataset}
}
\value{
list containing 1)a dataframe with correlation coefficients (CC), p-values and FDR-q values, 2) list of a) observed CC distribution and b)null CC distribution
}
\description{
SCA

Data for SCA
}
