% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_Cluster_Resolution_Harmony.R
\name{Test_Cluster_Resolution_Harmony}
\alias{Test_Cluster_Resolution_Harmony}
\title{A Test_Cluster_Resolution_Harmony Function}
\usage{
Test_Cluster_Resolution_Harmony(
  Temp.object,
  saveDIR,
  IdentToBatchCorrect = "orig.ident",
  ThetaToBatchCorrect = 2,
  SuffixName = "Testing_Cluster_Resolution",
  ColToPlot = ColToPlot,
  ColPaletteToPlot = ColPaletteToPlot,
  mingenes = 500,
  PCAuse = 25,
  ClusResList = c(0.1, 0.15, 0.2, 0.3),
  ClusOrder = ClusOrderFrom1,
  DownSamplePCA = 1000
)
}
\arguments{
\item{Temp.object}{Seurat objects to be used for the QC plots.}

\item{saveDIR}{Directory to save the plots.}

\item{IdentToBatchCorrect}{Identity to be used for Harmony batch correction}

\item{ThetaToBatchCorrect}{Theta values to be used while running harmony. Diversity clustering penalty parameter. Specify for each variable in group.by.vars. Default theta=2. theta=0 does not encourage any diversity. Larger values of theta result in more diverse clusters.}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{ColToPlot}{Mention one ident only}

\item{ColPaletteToPlot}{Mention one palette only for the ident}

\item{mingenes}{minimum gene number that will be mentioned in the output file name}

\item{PCAuse}{PCA dim to use}

\item{ClusResList}{Vector of the PCAs to be used for testing}

\item{DownSamplePCA}{#Cells to plot in UMAP for PCA testing}
}
\description{
This function allows you to compare different cluster resolutions.
}
\examples{
Test_Cluster_Resolution_Harmony()
}
\keyword{ClusResList}
\keyword{ColPaletteToPlot}
\keyword{ColToPlot,}
\keyword{IdentToBatchCorrect,}
\keyword{PCAuse}
\keyword{SuffixName,}
\keyword{Temp.object,}
\keyword{ThetaToBatchCorrect,}
\keyword{mingenes}
\keyword{saveDIR,}
