% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_PCA_dim_Harmony.R
\name{Test_PCA_dim_Harmony}
\alias{Test_PCA_dim_Harmony}
\title{A Test_PCA_dim_Harmony Function}
\usage{
Test_PCA_dim_Harmony(
  Temp.object,
  saveDIR,
  IdentToBatchCorrect = "orig.ident",
  ThetaToBatchCorrect = 2,
  SuffixName = "Testing_PCA_Dim",
  ColNamesToPlot = ColNamesToPlot,
  ColPaletteToPlot = ColPaletteToPlot,
  mingenes = 500,
  PCAdimList = c(20, 25, 30, 40, 50),
  ClusOrder = ClusOrderFrom1
)
}
\arguments{
\item{Temp.object}{Seurat objects to be used for the QC plots.}

\item{saveDIR}{Directory to save the plots.}

\item{IdentToBatchCorrect}{Identity to be used for Harmony batch correction (Max 1)}

\item{ThetaToBatchCorrect}{Theta values to be used while running harmony. Diversity clustering penalty parameter. Specify for each variable in group.by.vars. Default theta=2. theta=0 does not encourage any diversity. Larger values of theta result in more diverse clusters.}

\item{SuffixName}{Suffix. to be added in the directory name as}

\item{ColNamesToPlot}{Vector of identities to plot (Max = 3)}

\item{ColPaletteToPlot}{list of color palettes to plot for the identities (Max = 3)}

\item{mingenes}{minimum gene number that will be mentioned in the output file name}

\item{PCAdimList}{Vector of the PCAs to be used for testing}
}
\description{
This function  allows you to compare different no. of PCAs for Harmony. Cluster Resolution is set to 0.1.
}
\examples{
Test_PCA_dim_Harmony()
}
\keyword{ColNamesToPlot,}
\keyword{ColPaletteToPlot}
\keyword{IdentToBatchCorrect,}
\keyword{PCAdimList}
\keyword{SuffixName,}
\keyword{Temp.object,}
\keyword{ThetaToBatchCorrect,}
\keyword{mingenes}
\keyword{saveDIR,}
