% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA_data.R
\docType{data}
\name{vanGalen_HSPC}
\alias{vanGalen_HSPC}
\title{van Galen et al (2019) normal HSPC data (BM5 CD34p): cell types annotated by clustering and cell type specific marker gene expressions}
\format{
SCE objects of normal HSPC datasets
\describe{
       \item{assays}{counts: UMI counts single cell gene expression matrix & logcounts: log2 counts scGEP matrix}
       \item{colData}{DataFrame of single cell meta data. "SCA_cls" slot contains cell types annotated by original authors}
       }
       @source{add source later}
       @examples
       # example code
       data(vanGalen_HSPC)
}
\usage{
vanGalen_HSPC
}
\description{
van Galen et al (2019) normal HSPC data (BM5 CD34p): cell types annotated by clustering and cell type specific marker gene expressions
}
\keyword{datasets}
