% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA_data.R
\docType{data}
\name{Velten_HSPC}
\alias{Velten_HSPC}
\title{Velten et al (2017) normal HSPC data: index-sorted scRNAseq data}
\format{
SCE objects of normal HSPC datasets
\describe{
       \item{assays}{counts: read counts single cell gene expression matrix & logcounts: log2 counts scGEP matrix}
       \item{colData}{DataFrame of single cell meta data. "SCA_cls" slot contains cleaned up version of cell types annotated by original authors}
       }
       @source{add source later}
       @examples
       # example code
       data(Velten_HSPC)
}
\usage{
Velten_HSPC
}
\description{
Velten et al (2017) normal HSPC data: index-sorted scRNAseq data
}
\keyword{datasets}
